+++ CHANGELOG.rst
@@ -4,9 +4,10 @@
 Changelog
 =============================
 
-Version 0.7.2 - (2010-Sep-06)
+Version 0.8.a1 (2010-Sep-07)
 ==============================
- * Added ``form_template`` in :class:`djpcms.views.appsite.ModelApplication`` for customizing uniform rendering.
+ * removed a couple of obsolete function in :class:`djpcms.views.response.DjpResponse` and added the new function ``instancecode`` to return an unique code for an instance of a model.
+ * Added ``form_template`` in :class:`djpcms.views.appsite.ModelApplication` for customizing uniform rendering.
  * updated to jQuery 1.4.2 and added two jQuery UI themes.
  * Added :setting:`DJPCMS_STYLE` setting for specifying ``css`` style.
  * :class:`djpcms.views.cache.PageCache` works when django sessions are not available. Previously it was failing.